requirements.txt
pandas
openpyxl
pillow
google-generativeai
python-dotenv
